fix: Reorder mutebot checks in event handlers

+++ events/handler.js
@@ -148,8 +148,6 @@ module.exports = (bot) => {
         const groupDb = await db.get(`group.${groupId}`) || {};
 
         // Pengecekan mode bot (group, private, self)
-        if (groupDb?.mutebot === true && !isOwner && !isAdmin) return;
-        if (groupDb?.mutebot === "owner" && !isOwner) return;
         if (botDb?.mode === "group" && isPrivate && !isOwner && !userDb?.premium) return;
         if (botDb?.mode === "private" && isGroup && !isOwner && !userDb?.premium) return;
         if (botDb?.mode === "self" && !isOwner) return;
@@ -160,6 +158,8 @@ module.exports = (bot) => {
         if (hour >= 0 && hour < 6 && !isOwner && !userDb?.premium) return;
 
         // Pengecekan mute pada grup
+        if (groupDb?.mutebot === true && !isOwner && !isAdmin) return;
+        if (groupDb?.mutebot === "owner" && !isOwner) return;
         const muteList = groupDb?.mute || [];
         if (muteList.includes(senderId)) await ctx.deleteMessage(m.key);
